use new pattern for UsersReducer
